add overview of Vue ecosystem
I originally posted this in a comment to [Issue #3](https://github.com/farmOS/farmOS-client/issues/3#issuecomment-385042481). It should still be cleaned up and expanded, but I feel it hits all the major beats of how we're leveraging the Vue ecosystem for the client and data plugin. The login plugin should be addressed, too, but it's pretty similar to the data plugin; it's also in need of some overhaul so I'd like to wait til that's done.