Update Matrix chat links
Use generic Matrix links instead of "riot.im" for the chat rooms.

The Matrix app "Riot" has recently been re-branded to "Element" and following a "riot.im" link prompts the user to move to "element.io". Let's be client-agnostic instead, using "matrix.to" links.

(The "matrix.to" site is currently undisputedly ugly. However, (a) work is under way to improve it, and (b) some Matrix apps recognize matrix.to URLs and directly open the room in their already installed app bypassing that page.)